requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-filter==24.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
idna==3.6
Markdown==3.6
oauthlib==3.2.2
phonenumbers==8.13.32
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
